requirements.txt
# Market Dashboard Requirements
# Phase 1 Implementation

# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0

# Data fetching
yfinance>=0.2.28
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database

# Environment management
python-dotenv>=1.0.0

# Optional but recommended
scipy>=1.11.0  # For advanced calculations
